% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BondUtils.R
\name{BondAC}
\alias{BondAC}
\title{Bond accrued interest}
\usage{
BondAC(bond, dtSettlement)
}
\arguments{
\item{bond}{data structure generated by \code{Bond}}

\item{dtSettlement}{settlement date}
}
\value{
the accrued interest
}
\description{
This function computes the accrued interest on a bond by the formula
\deqn{
ac = c N \frac{m}{365}
}
}
\details{
with
\describe{
\item{\eqn{c}}{coupon rate}
\item{\eqn{N}}{nominal amount}
\item{\eqn{m}}{actual number of days since last coupon payment}
}
}
\examples{
b374 <- Bond('374', myDate('01/01/2000'),
             myDate('04/01/2019'), .0375, 100, 'a')
a <- BondAC(b374, myDate('03/01/2012'))
}
