% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BondUtils.R
\name{BondCvx}
\alias{BondCvx}
\title{Bond convexity}
\usage{
BondCvx(bond, dtSettlement, yield, price = NULL)
}
\arguments{
\item{bond}{data structure generated by \code{Bond}}

\item{dtSettlement}{settlement date}

\item{yield}{(optional) bond yield}

\item{price}{(optional) bond price}
}
\value{
Bond convexity
}
\description{
This function computes the convexity of a bond by the formula
\deqn{
Cx = \frac{\partial^2 P(y)}{\partial y^2}
}
given the bond price or yield.
}
\details{
with
\describe{
\item{\eqn{P(y)}}{bond price, function of yield}
}
}
\examples{
b374 <- Bond('374', myDate('01/01/2000'),
             myDate('04/01/2019'), .0375, 100, 'a')
Cvx <- BondCvx(b374, myDate('03/01/2012'), yield=.04)
}
