\name{BondPV01}
\alias{BondPV01}
\title{Bond PV01}
\usage{
  BondPV01(bond, dtSettlement, yield, price = NULL)
}
\arguments{
  \item{bond}{data structure generated by \code{Bond}}

  \item{dtSettlement}{settlement date}

  \item{yield}{(optional) bond yield}

  \item{price}{(optional) bond price}
}
\value{
  PV01
}
\description{
  This function computes the PV01 (present value of one
  basis point) of a bond by the formula \deqn{ PV01 =
  \frac{1}{10000} \frac{\partial P(y)}{\partial y} } given
  the bond price or yield.
}
\details{
  with \describe{ \item{\eqn{P(y)}}{bond price, function of
  yield} }
}
\examples{
b374 <- Bond('374', myDate('01/01/2000'),
             myDate('04/01/2019'), .0375, 100, 'a')
pv01 <- BondPV01(b374, myDate('03/01/2012'), yield=.04)
}

