% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DateUtils.r
\name{myDate}
\alias{myDate}
\title{Date constructor}
\usage{
myDate(dt)
}
\value{
a Date object
}
\description{
Date creation
}
\details{
A shortcut for date creation, locale independent
validates day and month ranges
}
\examples{
d1 <- myDate('30-10-2010')
d2 <- myDate('30/10/2010')
d3 <- myDate('30.10.2010')
d4 <- myDate('30oct2010')
(d1 == d2) & (d1 == d3) & (d1 == d4)
}
