% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BondUtils.R
\name{Price2BondYield}
\alias{Price2BondYield}
\title{Bond yield from price}
\usage{
Price2BondYield(bond, dtSettlement, price)
}
\arguments{
\item{bond}{data structure generated by \code{Bond}}

\item{dtSettlement}{settlement date}

\item{price}{bond dirty price}
}
\value{
bond yield
}
\description{
This function computes the bond yield, given its dirty price by solving for \eqn{y} the equation:
\deqn{
P = \sum_{i=1}^n F_i (1+y)^{-\frac{d_i - d_0}{365}}
}
with:
\describe{
\item{\eqn{F_i}}{Cash flow paid at date \eqn{d_i}}
\item{\eqn{y}}{yield}
\item{\eqn{d_i}}{Cash flow date, measured in days}
\item{\eqn{d_0}}{settlement date}
}
}
\examples{
b374 <- Bond('374', myDate('01/01/2000'),
             myDate('04/01/2019'), .0375, 100, 'a')
y <- Price2BondYield(b374, myDate('03/01/2012'),
     price=101)
}
