\name{TrinomialTreePlot}
\alias{TrinomialTreePlot}
\title{Trinomial tree plot}
\usage{
  TrinomialTreePlot(TrinomialTreeValues, dx = -0.025,
    dy = 0.4, cex = 1, digits = 2, ...)
}
\arguments{
  \item{TrinomialTreeValues}{tree values grid}

  \item{dx}{step}

  \item{digits}{format}
}
\description{
  Trinomial tree plot
}

