% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DateUtils.r
\name{tDiff}
\alias{tDiff}
\title{time difference}
\usage{
tDiff(dtFirst, dtLast)
}
\arguments{
\item{dtFirst}{a date or timeDate}

\item{dtLast}{a date or timeDate}
}
\value{
the time difference in year fraction
}
\description{
time difference in fraction of years, given 2 timeDate or Date objects
}
